Packaging item-related classes
