+++ src/main/java/seedu/address/logic/parser/AddressBookParser.java
@@ -8,19 +8,19 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import seedu.address.commons.core.LogsCenter;
-import seedu.address.logic.commands.AddItemCommand;
 
 import seedu.address.logic.commands.ClearCommand;
 import seedu.address.logic.commands.Command;
-import seedu.address.logic.commands.DecreaseItemCommand;
-import seedu.address.logic.commands.DeleteItemCommand;
-import seedu.address.logic.commands.EditItemCommand;
 import seedu.address.logic.commands.ExitCommand;
-import seedu.address.logic.commands.FindItemCommand;
 import seedu.address.logic.commands.HelpCommand;
 import seedu.address.logic.commands.HistoryCommand;
-import seedu.address.logic.commands.IncreaseItemCommand;
-import seedu.address.logic.commands.ListItemCommand;
+import seedu.address.logic.commands.ItemCommand.AddItemCommand;
+import seedu.address.logic.commands.ItemCommand.DecreaseItemCommand;
+import seedu.address.logic.commands.ItemCommand.DeleteItemCommand;
+import seedu.address.logic.commands.ItemCommand.EditItemCommand;
+import seedu.address.logic.commands.ItemCommand.FindItemCommand;
+import seedu.address.logic.commands.ItemCommand.IncreaseItemCommand;
+import seedu.address.logic.commands.ItemCommand.ListItemCommand;
 import seedu.address.logic.commands.MemberCommand.AddMemberCommand;
 import seedu.address.logic.commands.MemberCommand.DeleteMemberCommand;
 import seedu.address.logic.commands.MemberCommand.EditMemberCommand;
@@ -43,6 +43,12 @@ import seedu.address.logic.commands.ledger.CreditCommand;
 import seedu.address.logic.commands.ledger.DebitCommand;
 import seedu.address.logic.commands.ledger.DeleteLedgerCommand;
 import seedu.address.logic.commands.ledger.OpenLedgerCommand;
+import seedu.address.logic.parser.Item.AddItemCommandParser;
+import seedu.address.logic.parser.Item.DecreaseItemCommandParser;
+import seedu.address.logic.parser.Item.DeleteItemCommandParser;
+import seedu.address.logic.parser.Item.EditItemCommandParser;
+import seedu.address.logic.parser.Item.FindItemCommandParser;
+import seedu.address.logic.parser.Item.IncreaseItemCommandParser;
 import seedu.address.logic.parser.Member.AddCommandParser;
 import seedu.address.logic.parser.Member.DeleteCommandParser;
 import seedu.address.logic.parser.Member.EditCommandParser;
