+++ src/main/java/seedu/address/logic/parser/Item/EditItemCommandParser.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.Item;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
@@ -7,8 +7,12 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_ITEM_NAME;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_ITEM_QUANTITY;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.EditItemCommand;
-import seedu.address.logic.commands.EditItemCommand.EditItemDescriptor;
+import seedu.address.logic.commands.ItemCommand.EditItemCommand;
+import seedu.address.logic.commands.ItemCommand.EditItemCommand.EditItemDescriptor;
+import seedu.address.logic.parser.ArgumentMultimap;
+import seedu.address.logic.parser.ArgumentTokenizer;
+import seedu.address.logic.parser.Parser;
+import seedu.address.logic.parser.ParserUtil;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 /**
