+++ src/main/java/seedu/address/logic/parser/Item/FindItemCommandParser.java
@@ -1,10 +1,11 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.Item;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import java.util.Arrays;
 
-import seedu.address.logic.commands.FindItemCommand;
+import seedu.address.logic.commands.ItemCommand.FindItemCommand;
+import seedu.address.logic.parser.Parser;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.item.ItemNameContainsKeywordsPredicate;
